fix: correct typo and rename prop in error.tsx and ErrorPage.tsx
