+++ frontend/src/app/customers/[id]/error.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ErrorPage from "@/app/components/elements/erroPage/ErrorPage";
+import ErrorPage from "@/app/components/elements/errorPage/ErrorPage";
 import { useLanguage } from "@/app/contexts/LanguageContext";
 
 export default function Error({
@@ -11,10 +11,14 @@ export default function Error({
   reset: () => void;
 }) {
   const { language } = useLanguage();
-  const [messageJp, messageEn] = error.message.split(" / ");
-  const errorMessages = { messageJp, messageEn };
+  const [messageEn, messageJa] = error.message.split(" / ");
+  const errorMessages = { messageEn, messageJa: messageJa || messageEn };
 
   return (
-    <ErrorPage reset={reset} errorMessages={errorMessages} lng={language} />
+    <ErrorPage
+      reset={reset}
+      errorMessages={errorMessages}
+      language={language}
+    />
   );
 }
